+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Add support for the `GL_OES_element_index_uint` extension.
  - Fixed OpenGL ES 3.2 not working.
  - Add support for `IntVec{2|3|4}` and `UnsignedIntVec{2|3|4}` uniform types.
+ - Add a work-around for Radeon drivers crashing with 32+ texture units.
 
 ## Version 0.8.5 (2015-08-12)